+++ wubiq-common/src/net/sf/wubiq/common/PropertyKeys.java
@@ -13,8 +13,12 @@ public final class PropertyKeys {
 
 	public static final String WUBIQ_CLIENT_FORCE_SERIALIZED_CONNECTION = "wubiq.client.force.serialized.communication";
 	public static final String WUBIQ_CLIENT_CONNECTION_RETRIES = "wubiq.client.connection.retries";
+	public static final String WUBIQ_CLIENT_KEEP_ALIVE = "wubiq.client.keep_alive";
+	public static final String WUBIQ_CLIENT_SUPPRESS_LOGS = "wubiq.client.suppress_logs";
+
 	public static final String WUBIQ_FONTS_DOTMATRIX_DEFAULT = "wubiq.fonts.dotmatrix.default";
 	public static final String WUBIQ_FONTS_DOTMATRIX_FORCE_LOGICAL = "wubiq.fonts.dotmatrix.force.logical";
+
 	public static final String WUBIQ_PRINTERS_PHOTO = "wubiq.printers.photo";
 	public static final String WUBIQ_PRINTERS_DOTMATRIX = "wubiq.printers.dotmatrix";
 	public static final String WUBIQ_PRINTERS_DOTMATRIX_HQ = "wubiq.printers.dotmatrix.hq";